requirements/requirements-dev.txt
transformers==4.57.3
albumentations==2.0.o
torch==2.9.1
torchvision==0.24.1
matplotlib==3.10.8
supervision==0.27.0
hydra-core==1.3.2
clearml==2.1.2
tensorboard==2.20.0